refactor: refactored input types
