+++ src/routes/graphql/post/inputPost.ts
@@ -19,18 +19,4 @@ export const ChangePostInput = new GraphQLInputObjectType({
   }),
 });
 
-export interface CreatePostInputType {
-  dto: {
-    title: string;
-    content: string;
-    authorId: string;
-  };
-}
 
-export interface ChangePostInputType {
-  id: string;
-  dto: {
-    title: string;
-    content: string;
-  };
-}
