+++ src/routes/graphql/user/inputUser.ts
@@ -13,17 +13,3 @@ export const ChangeUserInput = new GraphQLInputObjectType({
   fields: () => ({ name: { type: GraphQLString }, balance: { type: GraphQLFloat } }),
 });
 
-export interface CreateUserInputType {
-  dto: {
-    name: string;
-    balance: number;
-  };
-}
-
-export interface ChangeUserInputType {
-  id: string;
-  dto: {
-    name: string;
-    balance: number;
-  };
-}
